+++ src/index.js
@@ -34,7 +34,6 @@ export * from './components/RadioButton';
 export * from './components/ResizeSensor';
 export * from './components/Scrollbar'; // deprecated
 export * from './components/SectionPreloader';
-export * from './components/Select';
 export * from './components/SplashError';
 export * from './components/SplashModal';
 export * from './components/Spin';